// src/es/schemas/fleet.search.ts
/*
 * Copyright Elasticsearch B.V. and contributors
 * SPDX-License-Identifier: Apache-2.0
 */

/* eslint-disable @typescript-eslint/no-redeclare */
import { z } from 'zod'

import { AggregateName, ClusterStatistics, Duration, ExpandWildcards, Field, Fields, Id, IndexAlias, IndexName, Routing, ScriptField, ScrollId, SearchType, ShardStatistics, SlicedScroll, Sort, SortResults, SuggestMode, SuggestionName, double, integer, long } from './_types.js'
import { AggregationsAggregate, AggregationsAggregationContainer } from './_types.aggregations.js'
import { MappingRuntimeFields } from './_types.mapping.js'
import { QueryDslFieldAndFormat, QueryDslOperator, QueryDslQueryContainer } from './_types.query_dsl.js'
import { FleetCheckpoint } from './fleet.js'
import { SearchFieldCollapse, SearchHighlight, SearchHitsMetadata, SearchPointInTimeReference, SearchProfile, SearchRescore, SearchSourceConfig, SearchSuggest, SearchSuggester, SearchTrackHits } from './search.js'

/**
 * Run a Fleet search.
 *
 * The purpose of the Fleet search API is to provide an API where the search will be run only
 * after the provided checkpoint has been processed and is visible for searches inside of Elasticsearch.
 */
export const FleetSearchRequest = z.object({
  index: z.union([z.lazy(() => IndexName), IndexAlias]).describe('A single target to search. If the target is an index alias, it must resolve to a single index.').meta({ found_in: 'path' }),
  allow_no_indices: z.boolean().describe('A setting that does two separate checks on the index expression. If `false`, the request returns an error (1) if any wildcard expression (including `_all` and `*`) resolves to zero matching indices or (2) if the complete set of resolved indices, aliases or data streams is empty after all expressions are evaluated. If `true`, index expressions that resolve to no indices are allowed and the request returns an empty result.').optional().meta({ found_in: 'query' }),
  analyzer: z.string().optional().meta({ found_in: 'query' }),
  analyze_wildcard: z.boolean().optional().meta({ found_in: 'query' }),
  batched_reduce_size: z.lazy(() => long).optional().meta({ found_in: 'query' }),
  ccs_minimize_roundtrips: z.boolean().optional().meta({ found_in: 'query' }),
  default_operator: z.lazy(() => QueryDslOperator).optional().meta({ found_in: 'query' }),
  df: z.string().optional().meta({ found_in: 'query' }),
  expand_wildcards: z.lazy(() => ExpandWildcards).optional().meta({ found_in: 'query' }),
  ignore_throttled: z.boolean().optional().meta({ found_in: 'query' }),
  ignore_unavailable: z.boolean().describe('If `false`, the request returns an error if it targets a concrete (non-wildcarded) index, alias, or data stream that is missing, closed, or otherwise unavailable. If `true`, unavailable concrete targets are silently ignored.').optional().meta({ found_in: 'query' }),
  lenient: z.boolean().optional().meta({ found_in: 'query' }),
  max_concurrent_shard_requests: z.lazy(() => integer).optional().meta({ found_in: 'query' }),
  preference: z.string().optional().meta({ found_in: 'query' }),
  pre_filter_shard_size: z.lazy(() => long).optional().meta({ found_in: 'query' }),
  request_cache: z.boolean().optional().meta({ found_in: 'query' }),
  routing: z.lazy(() => Routing).optional().meta({ found_in: 'query' }),
  scroll: z.lazy(() => Duration).optional().meta({ found_in: 'query' }),
  search_type: z.lazy(() => SearchType).optional().meta({ found_in: 'query' }),
  suggest_field: z.lazy(() => Field).describe('Specifies which field to use for suggestions.').optional().meta({ found_in: 'query' }),
  suggest_mode: z.lazy(() => SuggestMode).optional().meta({ found_in: 'query' }),
  suggest_size: z.lazy(() => long).optional().meta({ found_in: 'query' }),
  suggest_text: z.string().describe('The source text for which the suggestions should be returned.').optional().meta({ found_in: 'query' }),
  typed_keys: z.boolean().optional().meta({ found_in: 'query' }),
  rest_total_hits_as_int: z.boolean().optional().meta({ found_in: 'query' }),
  _source_excludes: z.lazy(() => Fields).optional().meta({ found_in: 'query' }),
  _source_includes: z.lazy(() => Fields).optional().meta({ found_in: 'query' }),
  q: z.string().optional().meta({ found_in: 'query' }),
  wait_for_checkpoints: z.array(FleetCheckpoint).describe('A comma separated list of checkpoints. When configured, the search API will only be executed on a shard after the relevant checkpoint has become visible for search. Defaults to an empty list which will cause Elasticsearch to immediately execute the search.').optional().meta({ found_in: 'query' }),
  allow_partial_search_results: z.boolean().describe('If true, returns partial results if there are shard request timeouts or shard failures. If false, returns an error with no partial results. Defaults to the configured cluster setting `search.default_allow_partial_results`, which is true by default.').optional().meta({ found_in: 'query' }),
  aggregations: z.record(z.string(), z.lazy(() => AggregationsAggregationContainer)).optional().meta({ found_in: 'body' }),
  aggs: z.record(z.string(), z.lazy(() => AggregationsAggregationContainer)).optional(),
  collapse: z.lazy(() => SearchFieldCollapse).optional().meta({ found_in: 'body' }),
  explain: z.boolean().describe('If true, returns detailed information about score computation as part of a hit.').optional().meta({ found_in: 'body' }),
  ext: z.record(z.string(), z.any()).describe('Configuration of search extensions defined by Elasticsearch plugins.').optional().meta({ found_in: 'body' }),
  from: z.lazy(() => integer).describe('Starting document offset. By default, you cannot page through more than 10,000 hits using the from and size parameters. To page through more hits, use the search_after parameter.').optional().meta({ found_in: 'body' }),
  highlight: z.lazy(() => SearchHighlight).optional().meta({ found_in: 'body' }),
  track_total_hits: SearchTrackHits.describe('Number of hits matching the query to count accurately. If true, the exact number of hits is returned at the cost of some performance. If false, the response does not include the total number of hits matching the query. Defaults to 10,000 hits.').optional().meta({ found_in: 'body' }),
  indices_boost: z.array(z.record(z.lazy(() => IndexName), z.lazy(() => double))).describe('Boosts the _score of documents from specified indices.').optional().meta({ found_in: 'body' }),
  docvalue_fields: z.array(z.union([z.lazy(() => QueryDslFieldAndFormat), z.lazy(() => Field)])).describe('Array of wildcard (*) patterns. The request returns doc values for field names matching these patterns in the hits.fields property of the response.').optional().meta({ found_in: 'body' }),
  min_score: z.lazy(() => double).describe('Minimum _score for matching documents. Documents with a lower _score are not included in search results and results collected by aggregations.').optional().meta({ found_in: 'body' }),
  post_filter: z.lazy(() => QueryDslQueryContainer).optional().meta({ found_in: 'body' }),
  profile: z.boolean().optional().meta({ found_in: 'body' }),
  query: z.lazy(() => QueryDslQueryContainer).describe('Defines the search definition using the Query DSL.').optional().meta({ found_in: 'body' }),
  rescore: z.union([z.lazy(() => SearchRescore), z.array(z.lazy(() => SearchRescore))]).optional().meta({ found_in: 'body' }),
  script_fields: z.record(z.string(), z.lazy(() => ScriptField)).describe('Retrieve a script evaluation (based on different fields) for each hit.').optional().meta({ found_in: 'body' }),
  search_after: z.lazy(() => SortResults).optional().meta({ found_in: 'body' }),
  size: z.lazy(() => integer).describe('The number of hits to return. By default, you cannot page through more than 10,000 hits using the from and size parameters. To page through more hits, use the search_after parameter.').optional().meta({ found_in: 'body' }),
  slice: z.lazy(() => SlicedScroll).optional().meta({ found_in: 'body' }),
  sort: z.lazy(() => Sort).optional().meta({ found_in: 'body' }),
  _source: z.lazy(() => SearchSourceConfig).describe('Indicates which source fields are returned for matching documents. These fields are returned in the hits._source property of the search response.').optional().meta({ found_in: 'body' }),
  fields: z.array(z.union([z.lazy(() => QueryDslFieldAndFormat), z.lazy(() => Field)])).describe('Array of wildcard (*) patterns. The request returns values for field names matching these patterns in the hits.fields property of the response.').optional().meta({ found_in: 'body' }),
  suggest: SearchSuggester.optional().meta({ found_in: 'body' }),
  terminate_after: z.lazy(() => long).describe('Maximum number of documents to collect for each shard. If a query reaches this limit, Elasticsearch terminates the query early. Elasticsearch collects documents before sorting. Defaults to 0, which does not terminate query execution early.').optional().meta({ found_in: 'body' }),
  timeout: z.string().describe('Specifies the period of time to wait for a response from each shard. If no response is received before the timeout expires, the request fails and returns an error. Defaults to no timeout.').optional().meta({ found_in: 'body' }),
  track_scores: z.boolean().describe('If true, calculate and return document scores, even if the scores are not used for sorting.').optional().meta({ found_in: 'body' }),
  version: z.boolean().describe('If true, returns document version as part of a hit.').optional().meta({ found_in: 'body' }),
  seq_no_primary_term: z.boolean().describe('If true, returns sequence number and primary term of the last modification of each hit. See Optimistic concurrency control.').optional().meta({ found_in: 'body' }),
  stored_fields: z.lazy(() => Fields).describe('List of stored fields to return as part of a hit. If no fields are specified, no stored fields are included in the response. If this field is specified, the _source parameter defaults to false. You can pass _source: true to return both source fields and stored fields in the search response.').optional().meta({ found_in: 'body' }),
  pit: SearchPointInTimeReference.describe('Limits the search to a point in time (PIT). If you provide a PIT, you cannot specify an <index> in the request path.').optional().meta({ found_in: 'body' }),
  runtime_mappings: z.lazy(() => MappingRuntimeFields).describe('Defines one or more runtime fields in the search request. These fields take precedence over mapped fields with the same name.').optional().meta({ found_in: 'body' }),
  stats: z.array(z.string()).describe('Stats groups to associate with the search. Each group maintains a statistics aggregation for its associated searches. You can retrieve these stats using the indices stats API.').optional().meta({ found_in: 'body' })
}).meta({ id: 'FleetSearchRequest' })
export type FleetSearchRequest = z.infer<typeof FleetSearchRequest>

export const FleetSearchResponse = z.object({
  took: z.lazy(() => long),
  timed_out: z.boolean(),
  _shards: z.lazy(() => ShardStatistics),
  hits: z.lazy(() => SearchHitsMetadata),
  aggregations: z.record(z.lazy(() => AggregateName), AggregationsAggregate).optional(),
  _clusters: z.lazy(() => ClusterStatistics).optional(),
  fields: z.record(z.string(), z.any()).optional(),
  max_score: z.lazy(() => double).optional(),
  num_reduce_phases: z.lazy(() => long).optional(),
  profile: SearchProfile.optional(),
  pit_id: z.lazy(() => Id).optional(),
  _scroll_id: z.lazy(() => ScrollId).optional(),
  suggest: z.record(z.lazy(() => SuggestionName), z.array(SearchSuggest)).optional(),
  terminated_early: z.boolean().optional()
}).meta({ id: 'FleetSearchResponse' })
export type FleetSearchResponse = z.infer<typeof FleetSearchResponse>
